refactor: extract context window token status
